Move package rdf to chunker/rdf.
