+++ chunker/rdf/state.go
